Чуток подкорректировал свой лист
