+++ src/main/Main.java
@@ -1,3 +1,5 @@
+package main;
+
 public class Main {
     public static void main(String[] args)
     {
