+++ src/main/StringList.java
@@ -1,3 +1,5 @@
+package main;
+
 public interface StringList
 {
     // Добавление элемента.
