+++ src/main/exertions/StringIsNullException.java
@@ -1,0 +1,9 @@
+package main.exertions;
+
+public class StringIsNullException extends RuntimeException
+{
+    public StringIsNullException(String message)
+    {
+        super(message);
+    }
+}
